backend/requirements.txt
django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
neo4j==5.28.2
neomodel==5.5.0
python-dotenv==1.1.1
bcrypt==4.3.0
